+++ be/MicroserviziSpring/msmedico/src/main/java/com/msmedico/repository/InfermiereRepositiry.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface InfermiereRepositiry extends JpaRepository<Infermiere, Long> {
+public interface InfermiereRepositiry extends JpaRepository<Infermiere, Integer> {
 }
