+++ be/MicroserviziSpring/mspazienti/src/main/java/com/mspazienti/repository/PazienteRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PazienteRepository extends JpaRepository<Paziente, Long> {
+public interface PazienteRepository extends JpaRepository<Paziente, Integer> {
 }
